refactor(station): station 패키지를 만들어 관련 클래스들을 옮겼다.
